[finishes #161697459] User sign in panel and some modification on homepage
Javascript code responsible for sign in panel responsiveness are
inclueded in homepage header section. Javascript code for drop
down header menu for mobile view was modified to synchronise the
drop down menu and sign in panel, so that one close when other is
open.

Some html and css style was also modified and broken down to
make it more modular and easy to use, but this modification does
not introduce any error.

The word `employ` was changed to `employed` Sub-headline in
upper page section.