[finishes #161697811] User sign up page plus additional improvement on homepage
Homepage's `sign up` and `open your account` link was set to
point to registration page after `open your account` button
was changed to link.

Some html and css style was also modified and broken down to
make it more modular and easy to use and more CSS was added
to `custom.css`.